+++ CHANGELOG.md
@@ -22,6 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - \[SDK\] The `resource_type` args now have the default value of `local` in task creation functions.
   The corresponding arguments are keyword-only now.
   (<https://github.com/opencv/cvat/pull/5502>)
+- \[Server API\] Added missing pagination or pagination parameters in
+  `/project/{id}/tasks`, `/tasks/{id}/jobs`, `/jobs/{id}/issues`,
+  `/jobs/{id}/commits`, `/issues/{id}/comments`
+  (<https://github.com/opencv/cvat/pull/5557>)
 
 ### Deprecated
 - TDB